+++ lib/backup/view/widgets/backups_list.dart
@@ -23,57 +23,54 @@ class BackupsList extends StatelessWidget {
             itemCount: state.availableBackups.length,
             itemBuilder: (context, index) {
               final file = state.availableBackups[index];
-              final date =
-              DateFormat('MMM-dd-yyyy hh:mm:ss aaa')
+              final date = DateFormat('MMM-dd-yyyy hh:mm:ss aaa')
                   .format(file.createdTime!.toLocal());
               return Card(
                 child: ListTile(
-                    title: Text(
-                      overflow: TextOverflow.ellipsis,
-                      '${file.name}'
-                          .split('.')
-                          .first,
-                      style: TextStyle(
-                          fontSize: 20.sp,
-                          fontWeight: FontWeight.w600),
-                    ),
-                    subtitle: Text('$date'),
-                    leading: IconButton(
-                      icon: Icon(Icons.highlight_remove,
-                          color: Theme
-                              .of(context)
-                              .colorScheme
-                              .error),
-                      onPressed: () async {
-                        final success = await context
-                            .read<BackupCubit>()
-                            .deleteBackup(file);
-                        _showSnackbar(
-                            context,
-                            success,
-                            success ? 'Success !' : 'Ups !',
-                            success
-                                ? 'Backup has been deleted.'
-                                : 'Something went wrong.');
-                      },
-                    ),
-                    trailing: CircleAvatar(
-                      backgroundColor:
-                      colors.primaryColor[100],
-                      child: Icon(Icons.download),
-                    ),
-                    onTap: () async {
-                      final success = await context
-                          .read<BackupCubit>()
-                          .restoreBackup(file.id!);
+                  title: Text(
+                    overflow: TextOverflow.ellipsis,
+                    '${file.name}'.split('.').first,
+                    style:
+                        TextStyle(fontSize: 20.sp, fontWeight: FontWeight.w600),
+                  ),
+                  subtitle: Text('$date'),
+                  leading: IconButton(
+                    icon: Icon(Icons.highlight_remove,
+                        color: Theme.of(context).colorScheme.error),
+                    onPressed: () async {
+                      final success =
+                          await context.read<BackupCubit>().deleteBackup(file);
                       _showSnackbar(
                           context,
                           success,
                           success ? 'Success !' : 'Ups !',
                           success
-                              ? 'Backup has been restored.'
+                              ? 'Backup has been deleted.'
                               : 'Something went wrong.');
-                    }),
+                    },
+                  ),
+                  trailing: CircleAvatar(
+                    backgroundColor: colors.primaryColor[100],
+                    child: Icon(Icons.download),
+                  ),
+                  onTap: () async {
+                    await showConfirmRestoreDialog(
+                      context,
+                      () async {
+                        final success = await context
+                            .read<BackupCubit>()
+                            .restoreBackup(file.id!);
+                        _showSnackbar(
+                            context,
+                            success,
+                            success ? 'Success !' : 'Ups !',
+                            success
+                                ? 'Backup has been restored.'
+                                : 'Something went wrong.');
+                      },
+                    );
+                  },
+                ),
               );
             },
           );