+++ lib/backup/view/widgets/widgets.dart
@@ -2,3 +2,4 @@ export 'backup_button.dart';
 export 'backups_list.dart';
 export 'sign_in_button.dart';
 export 'google_profile.dart';
+export 'confirm_dialog.dart';