+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [4.?.?] - ??-??-????
 ### Added
 - Auto backup to Google Drive
+- Restored backup overrides live database
 
 ## [4.3.0] - 03-22-2025
 ### Added